Alterada estrutura de pastas, adicionado comentarios e corrigidos pequenos problemas

+++ src/pacotes/utils/processados.py
@@ -1,13 +1,23 @@
-from nomes_arquivos import LISTA_NOMES_PROCESSADOS
+from src.pacotes.utils.nomes_arquivos import LISTA_NOMES_PROCESSADOS
 import os
-from logs import logger
+from src.pacotes.utils.logs import logger
 
 arquivo_lista_nomes_processados = LISTA_NOMES_PROCESSADOS
 
-with open(arquivo_lista_nomes_processados, 'w', encoding='utf-8') as arquivo:
-    arquivo.write('')
+#Cria o documento se ele não existir
+if not os.path.exists(arquivo_lista_nomes_processados):
+    with open(arquivo_lista_nomes_processados, 'w', encoding='utf-8') as arquivo:
+        arquivo.write('')
 
-def NomeProcessado(nome: str) -> bool:
+def nome_processado(nome: str) -> bool:
+    """
+    Verifica se o nome já foi processado ou não.
+
+    :param nome: Nome a ser verificado
+    :type nome: str
+    :return: Retorna True se o nome já tiver sido processado, retorna False caso ainda não tenha sido processado
+    :rtype: bool
+    """
     logger.debug(f'Iniciando "NomeProcessado" para o nome {nome}')
     lista_nomes_processados = []
     with open(arquivo_lista_nomes_processados, 'r', encoding='utf-8') as arquivo:
@@ -21,21 +31,26 @@ def NomeProcessado(nome: str) -> bool:
         lista_nomes_processados = [nome.strip("'") for nome in nomes]
         lista_nomes_processados.pop()
         if(nome in lista_nomes_processados):
-            logger.debug(f'Finalizando True "NomeProcessado" para o nome {nome}')
             return True
         else:
-            logger.debug(f'Finalizando False "NomeProcessado" para o nome {nome}')
             return False
         
     logger.debug(f'Finalizando False "NomeProcessado" para o nome {nome}')
     return False
     
-def AnotarNomeProcessado(nome: str) -> None:
-    logger.debug(f'Anotando "NomeProcessado" para o nome {nome}')
+def anotar_nome_processado(nome: str) -> None:
+    """
+    Escreve o nome informado no documento de nomes processados
+    
+    :param nome: Nome completo do candidato
+    :type nome: str
+    """
+
+    logger.info(f'Anotando "NomeProcessado" para o nome {nome}')
     with open(arquivo_lista_nomes_processados, 'a', encoding='utf-8') as arquivo:
         arquivo.write(f'{nome}, ')
 
 if __name__=='__main__':
     lista_nomes = ['ALEXANDRE ANDRADE DOS ANJOS JACOME', 'ALINE FERNANDES DE SA FINTELMAM', 'ANDREA GADELHA NOBREGA LINS', 'BRENO SERIQUE NEVES', 'BRUNA XAVIER REZENDE', 'BRUNO MARCONDES KOZLOWSKI', 'SAYONARA EPIFANIO DOS SANTOS PEREIRA E PEREIRA', 'SERGIO SAN GREGORIO FAVERO', 'TIAGO DALTOE', 'VINICIUS DE LEMOS SILVA']
     # for nome in lista_nomes:
-    #     AnotarNomeProcessado(nome)
+    #     anotar_nome_processado(nome)
